fix(types): accept full u64 hex literals; preserve pointer types in &expr at call sites
Two related typechecker regressions surfaced in the home-os kernel after
issues #21 and #22 landed.

#24 — Large u64 hex literals like `0xFFFF000000000000` were rejected.
The parser fell back to `@bitCast(parseInt(u64) -> i64)` for values
beyond `i64.max`, leaving the AST with a sign-extended negative i64
that `integerLiteralFits` correctly refused for an unsigned destination
(it required `value >= 0`). Fix: store IntegerLiteral.value as i128 so
the full unsigned 64-bit range fits positively. The parser now refuses
anything beyond u64.max with the existing "exceeds u64 range" error.
`integerLiteralFits` is widened to i128 and gains an upper bound check
for u64. Codegen / interpreter / comptime sites narrow back to i64 by
bit-pattern, preserving large unsigned masks intact.

#25 — `&array[i]` and `&struct.field` produced the value type at call
sites instead of `*ValueType`. The hint-aware unary path
(`inferUnaryExprWithHint`) propagated the call-site hint into the
operand and returned the operand's type unchanged — fine for `Neg` /
`Not` / `BitNot`, but wrong for `AddressOf`, which must wrap the
operand in a Reference. Fix: special-case `AddressOf` in the hint path
so it unwraps one Reference layer before propagating into the operand,
then re-wraps the operand type in a Reference on return.

Verified against the home-os kernel: arch/arm64/{mmu,gic400,gic600},
mm/slab, net/wireguard, sched/scheduler, security/kernel_user_split,
block/{readahead,request_merge}, container/bridge, crypto/sha256 all
typecheck cleanly for the previously-failing constructs.

Tests
- diagnostic snapshots: 21_u64_hex_literal_max (positive),
  22_u64_hex_literal_overflow (negative — u32 still rejects 40-bit
  literal), 23_address_of_array_index (positive), and
  24_address_of_struct_field (positive)
- `zig build test`: 46 passed (pre-existing `--listen=-` runner panic
  acknowledged)
- `zig build test-diagnostics`: 46 passed
- `zig build fuzz -- --seconds 1`: no crashes / timeouts

Closes #24
Closes #25

+++ packages/ast/src/ast.zig
@@ -249,14 +249,18 @@ pub const Node = struct {
 /// Integer literal expression.
 ///
 /// Represents a constant integer value in the source code.
-/// Currently supports 64-bit signed integers.
+/// Stored as i128 so the full unsigned 64-bit range is representable
+/// without bitcast tricks. Hex literals like `0xFFFF000000000000`
+/// (> i64.max) keep their natural positive value, which lets the
+/// type-checker decide signedness against the destination type
+/// rather than guessing from a sign-extended bit pattern.
 ///
-/// Example: `42`, `-17`, `0`
+/// Example: `42`, `-17`, `0`, `0xFFFFFFFFFFFFFFFF`
 pub const IntegerLiteral = struct {
     /// Base node metadata
     node: Node,
-    /// The integer value
-    value: i64,
+    /// The integer value (i128 to fit full u64 range positively)
+    value: i128,
     /// Optional type suffix (e.g., "i32", "u64")
     type_suffix: ?[]const u8 = null,
 
@@ -267,7 +271,7 @@ pub const IntegerLiteral = struct {
     ///   - loc: Source location
     ///
     /// Returns: Initialized IntegerLiteral
-    pub fn init(value: i64, loc: SourceLocation) IntegerLiteral {
+    pub fn init(value: i128, loc: SourceLocation) IntegerLiteral {
         return .{
             .node = .{ .type = .IntegerLiteral, .loc = loc },
             .value = value,
@@ -283,7 +287,7 @@ pub const IntegerLiteral = struct {
     ///   - loc: Source location
     ///
     /// Returns: Initialized IntegerLiteral with type suffix
-    pub fn initWithType(value: i64, type_suffix: ?[]const u8, loc: SourceLocation) IntegerLiteral {
+    pub fn initWithType(value: i128, type_suffix: ?[]const u8, loc: SourceLocation) IntegerLiteral {
         return .{
             .node = .{ .type = .IntegerLiteral, .loc = loc },
             .value = value,
